Refactor loop indices and condition checks for enhanced type safety
- Replaced `int` with `size_t` in various loops and condition checks across multiple files, improving compatibility with larger datasets and ensuring safer memory management.
- Updated handling of indices in `CMarkMgr`, `CEditView_Command_New`, and `CDlgFuncList` to utilize `size_t`, reinforcing consistent type usage.
- Documented type usage distinctions in AGENTS.md, providing clear guidelines for new code and enhancing overall type consistency across the codebase.

+++ sakura_core/macro/CIfObj.cpp
@@ -134,7 +134,7 @@ public:
 			if( pBstrName ){
 				*pBstrName = SysAllocString( m_sName.c_str() );
 			}
-		}else if( 0 <= memid && memid < (int)m_MethodsRef.size() ){
+		}else if( 0 <= memid && static_cast<size_t>(memid) < m_MethodsRef.size() ){
 			if( pBstrName ){
 				*pBstrName = SysAllocString( m_MethodsRef[memid].Name );
 			}
